moving back test_utils in tset_active_learning_utils for name conflict
